fix: try to fix sql error issues #79

+++ server/src/main/resources/db/migration/V20221215210734__init_metric_statimate.sql
@@ -3,16 +3,16 @@
 CREATE TABLE `metric_estimate`
 (
     `id`         BIGINT AUTO_INCREMENT PRIMARY KEY,
-    `system_id`  BIGINT         NOT NULL,
-    `cost`       DECIMAL(10, 2) NOT NULL,
-    `month`      INT            NOT NULL,
-    `people`     INT            NOT NULL,
-    `name`       VARCHAR(255)   NOT NULL,
-    `files`      INT            NOT NULL,
-    `lines`      INT            NOT NULL,
-    `blanks`     INT            NOT NULL,
-    `comment`    INT            NOT NULL,
-    `code`       INT            NOT NULL,
-    `complexity` INT            NOT NULL
+    `system_id`  BIGINT       NOT NULL,
+    `cost`       FLOAT        NOT NULL,
+    `month`      FLOAT        NOT NULL,
+    `people`     FLOAT        NOT NULL,
+    `name`       VARCHAR(255) NOT NULL,
+    `files`      INT          NOT NULL,
+    `lines`      INT          NOT NULL,
+    `blanks`     INT          NOT NULL,
+    `comment`    INT          NOT NULL,
+    `code`       INT          NOT NULL,
+    `complexity` INT          NOT NULL
 ) ENGINE = InnoDB;
